break down frontend structure into smaller components
