+++ app/src/App.js
@@ -1,32 +1,14 @@
 import * as React from "react";
 
-import { Routes, Route, Link } from "react-router-dom";
+import Sightings from "./components/Sightings";
 
-import Sightings from "./Sightings";
-
-const App = () => (
-  <main>
-    <nav>
-      <Link to="/">Home</Link> | <Link to="dashboard">Dashboard</Link>
-    </nav>
-    <Routes>
-      <Route path="/" element={<Home />} />
-      <Route path="/dashboard" element={<Dashboard />} />
-    </Routes>
-  </main>
-);
-
-const Home = () => (
-  <>
-    <h1>Animal Sighting Tracker</h1>
-    <Sightings />
-  </>
-);
-
-const Dashboard = () => (
-  <>
-    <h1>Dashboard</h1>
-  </>
-);
+const App = () => {
+  return (
+    <main>
+      <h1>Animal Sighting Tracker</h1>
+      <Sightings />
+    </main>
+  );
+};
 
 export default App;
