+++ app/src/components/Card.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Card({ individual }) {
+  return (
+    <div>
+      <h2>holder {individual}</h2>
+    </div>
+  );
+}
+
+export default Card;
